Move cmake check before fmt installation

+++ OS/Ubuntu/Ubuntu_packages.sh
@@ -60,32 +60,20 @@ function install_ubuntu_dev() {
     if [[ $ONPI = true && $PROJECT = "LRA" ]];then
         installation_info "cmake 3.22.4"
         if ! which cmake 2>/dev/null; then
-            # download and compile, will take about 1 hour
-            # cd $PARENT_PATH
-            # wget https://github.com/Kitware/CMake/releases/download/v3.22.4/cmake-3.22.4.tar.gz
-            # tar zxvf cmake-3.22.4.tar.gz
-            # cd cmake-3.22.4
-            # ./bootstrap
-            # make
-            # sudo make install
-            # installation_end "cmake 3.22.4"
-
             # install binary files directly
             # install path to /opt/cmake
             cd /opt
-            sudo wget https://github.com/Kitware/CMake/releases/download/v3.22.4/cmake-3.22.4-linux-aarch64.sh
-            sudo chmod +x cmake-3.22.4-linux-aarch64.sh
-            sudo ./cmake-3.22.4-linux-aarch64.sh
+            # sudo wget https://github.com/Kitware/CMake/releases/download/v3.22.4/cmake-3.22.4-linux-aarch64.sh
+            # sudo chmod +x cmake-3.22.4-linux-aarch64.sh
+
+            sudo wget https://github.com/Kitware/CMake/releases/download/v3.22.4/cmake-3.22.4-linux-aarch64.tar.gz
+            sudo tar zxvf cmake-3.22.4-linux-aarch64.tar.gz
+
             # link to usr/local/bin
-            sudo ln -s /opt/cmake-3.22.4-linux-aarch64.sh/bin/* usr/local/bin
+            sudo ln -s /opt/cmake-3.22.4-linux-aarch64/bin/* /usr/local/bin/
             cd $PARENT_PATH
-            if [ which cmake 2>/dev/null ];then
-                installation_end "cmake 3.22.4"
-            else
-                color_red "cmake 3.22.4 installation failed, please check it\n\n"
-                echo -e $color_word
-            fi
         else
+            cmake --version
             color_red "cmake already installed, this project required cmake 3.22.4.
                 \nPlease installed it manually or remove it first"
             echo -e $color_word
@@ -93,7 +81,7 @@ function install_ubuntu_dev() {
         fi
     else
         installation_info "cmake default"
-        sudo apt install cmake
+        sudo apt install cmake -y
         installation_end "cmake default"
     fi
 
@@ -128,7 +116,6 @@ function install_ubuntu_dev() {
     installation_end "tree"
 
     # add vscode
-
     installation_info "code normal"
     sudo apt install code
     installation_end "code normal"
@@ -140,37 +127,58 @@ function install_ubuntu_dev() {
 }
 
 function install_ubuntu_LRA_onpi() {
+    # change time zone
+    sudo timedatectl set-timezone Asia/Taipei
+
     # add code for aarch64 (ARM64)
     cd $PARENT_PATH
-    wget https://update.code.visualstudio.com/latest/linux-deb-arm64/stable -O code.deb
-    sudo dpkg -i code.deb
-    sudo apt --fix-broken install -y
-    sudo dpkg -i code.deb
+    installation_info "VS Code arm64 stable"
+    if [ ! -f ./code.deb ];then
+        wget https://update.code.visualstudio.com/latest/linux-deb-arm64/stable -O code.deb
+        sudo dpkg -i code.deb
+        sudo apt --fix-broken install -y
+        sudo dpkg -i code.deb
 
-    code --install-extension austin.code-gnu-global
-    code --install-extension cschlosser.doxdocgen
-    code --install-extension Gruntfuggly.todo-tree
-    code --install-extension jeff-hykin.better-cpp-syntax
-    code --install-extension ms-python.python
-    code --install-extension ms-python.vscode-pylance
-    code --install-extension ms-toolsai.jupyter
-    code --install-extension ms-toolsai.jupyter-keymap
-    code --install-extension ms-toolsai.jupyter-renderers
-    code --install-extension ms-vscode.cmake-tools
-    code --install-extension ms-vscode.cpptools
-    code --install-extension ms-vscode.cpptools-extension-pack
-    code --install-extension ms-vscode.cpptools-themes
+        code --install-extension austin.code-gnu-global
+        code --install-extension cschlosser.doxdocgen
+        code --install-extension Gruntfuggly.todo-tree
+        code --install-extension jeff-hykin.better-cpp-syntax
+        code --install-extension ms-python.python
+        code --install-extension ms-python.vscode-pylance
+        code --install-extension ms-toolsai.jupyter
+        code --install-extension ms-toolsai.jupyter-keymap
+        code --install-extension ms-toolsai.jupyter-renderers
+        code --install-extension ms-vscode.cmake-tools
+        code --install-extension ms-vscode.cpptools
+        code --install-extension ms-vscode.cpptools-extension-pack
+        code --install-extension ms-vscode.cpptools-themes
 
-    # maximum file watcher number of vscode (debug usage)
-    # ref: https://code.visualstudio.com/docs/setup/linux#_visual-studio-code-is-unable-to-watch-for-file-changes-in-this-large-workspace-error-enospc
-    echo 'fs.inotify.max_user_watches=524288' | sudo tee -a /etc/sysctl.conf
-    sudo sysctl -p
+        # maximum file watcher number of vscode (debug usage)
+        # ref: https://code.visualstudio.com/docs/setup/linux#_visual-studio-code-is-unable-to-watch-for-file-changes-in-this-large-workspace-error-enospc
+        if [ ! grep 'max_user_watches' /etc/sysctl.conf ];then
+            echo 'fs.inotify.max_user_watches=524288' | sudo tee -a /etc/sysctl.conf
+            sudo sysctl -p
+        fi
+    else
+        color_green "VS Code already downloaded"
+        echo -e $color_word
+    fi
+    installation_end "VS Code arm64 stable"
+
+    # confirm cmake state
+    if which cmake 2>/dev/null ;then
+        installation_end "cmake 3.22.4"
+    else
+        color_red "cmake 3.22.4 installation failed, please check it\n\n"
+        echo -e $color_word
+        sudo dpkg --configure -a
+    fi
 
     # install some libraries for project - LRA on Raspberry 4b
     # C++ fmt
     installation_info "fmt"
     # TODO:check fmt existed or not
-    if [[ ! -d "/usr/local/include/fmt" && ! -d "/usr/include/fmt" ]]
+    if [[ ! -d /usr/local/include/fmt && ! -d /usr/include/fmt ]]
     then
         cd $PARENT_PATH
         git clone https://github.com/fmtlib/fmt.git
@@ -188,15 +196,16 @@ function install_ubuntu_LRA_onpi() {
     # wiringPi
     installation_info "wiringPi"
     cd $PARENT_PATH
-    if find . -name wiringpi-2.61-g.deb ; then
-        echo "wiringpi-2.61-g.deb already downloaded"
+    if [ -f ./wiringpi-2.61-g.deb ] ; then
+        color_green "wiringpi-2.61-g.deb already downloaded"
+        echo -e $color_word
     else
     # Oops model 17 problem
         sudo dpkg --add-architecture armhf
         sudo apt update
         wget https://github.com/guation/WiringPi-arm64/releases/download/2.61-g/wiringpi-2.61-g.deb
     fi
-    sudo apt install -f ./wiringpi-*-g.deb
+    sudo apt install -f ./wiringpi-*-g.deb -y
 
     # gpio -readall group problem
     sudo apt install rpi.gpio-common
@@ -209,7 +218,7 @@ function install_ubuntu_LRA_onpi() {
     sudo apt install libi2c-dev -y
     # add $USER to i2c group
     sudo adduser $USER i2c
-    sudo su $USER
+    # sudo su $USER
     installation_end "libi2c-dev"
 
     # vcgencmd
@@ -220,7 +229,7 @@ function install_ubuntu_LRA_onpi() {
 
     # main program
     installation_info "LRA main program"
-    if [ ! -d "~/LRA_Raspberry4b" ];then
+    if [ ! -d ~/LRA_Raspberry4b ];then
         # clone from git
         cd ~
         git clone https://github.com/DennisLiu16/LRA_Raspberry4b.git
@@ -235,7 +244,7 @@ function install_ubuntu_LRA_onpi() {
         GROUP="spi"
         if [ ! $(getent group $GROUP) ]; then
             sudo groupadd $GROUP
-            sudo su $USER
+            # sudo su $USER
         fi
 
         if  id -nG "$USER" | grep -qw "$GROUP" ; then
@@ -243,14 +252,15 @@ function install_ubuntu_LRA_onpi() {
             echo "$USER was already in $GROUP"
         else
             sudo adduser $USER $GROUP
-            sudo su $USER
+            # sudo su $USER
         fi
 
         if grep '#for LRA SPI' /etc/udev/rules.d/50-spi.rules ; then
             echo "50-spi.rules for LRA already exists"
         else
             # avoid >> error
-            echo -e '\n\nset rule for LRA spi\n\n'
+            color_green '\n\nset rule for LRA spi\n\n'
+            echo -e $color_word
             echo -e "#for LRA SPI\nSUBSYSTEM==\"spidev\",KERNEL==\"spidev0.*\",GROUP=\"spi\",MODE=\"0660\"" | sudo tee -a /etc/udev/rules.d/50-spi.rules
         fi
 
@@ -264,7 +274,7 @@ function install_ubuntu_LRA_onpi() {
 
     # LRA web
     installation_info "LRA Web"
-    if [ ! -d "~/LRA_Web" ];then
+    if [ ! -d ~/LRA_Web ];then
         cd ~
         git clone https://github.com/DennisLiu16/LRA_Web.git
         # packages required
@@ -285,13 +295,18 @@ function install_ubuntu_LRA_onpi() {
     installation_end "LRA Web"
 
     # TODO:add global alias
-    cd ~
-    echo "alias cdc='cd ~/LRA_Raspberry4b'" >> .bash_aliases
-    echo "alias cdw='cd ~/LRA_Web'">> .bash_aliases
-    # wifi 
-    echo "alias getMAC='cat /sys/class/net/wlan0/address'" >> .bash_aliases
-    # measure
-    echo "alias temp='vcgencmd measure_temp'" >> .bash_aliases
+    if [ grep 'getMAC' ~/.bash_aliases ];then
+        color_green "Aliases already existed"
+        echo -e $color_word
+    else
+        cd ~
+        echo "alias cdc='cd ~/LRA_Raspberry4b'" >> .bash_aliases
+        echo "alias cdw='cd ~/LRA_Web'">> .bash_aliases
+        # wifi 
+        echo "alias getMAC='cat /sys/class/net/wlan0/address'" >> .bash_aliases
+        # measure
+        echo "alias temp='vcgencmd measure_temp'" >> .bash_aliases
+    fi
     cd $PARENT_PATH
 }
 
@@ -335,7 +350,7 @@ do
     ;;
     --reboot)
         ENABLE_REBOOT=true
-        shift 1
+        shift 
     ;;
     --onPi)
         ONPI=true
@@ -356,28 +371,28 @@ do
     esac
     done
 
-if [ $ONPI == true ];then
+if [ $ONPI = true ];then
     color_red $ONPI
     sONPI=$color_word
 else
     sONPI=$ONPI
 fi
 
-if [ $ENABLE_REBOOT == true ];then
+if [ $ENABLE_REBOOT = true ];then
     color_red $ENABLE_REBOOT
     sENABLE_REBOOT=$color_word
 else
     sENABLE_REBOOT=$ENABLE_REBOOT
 fi
 
-if [ ! $PROJECT == none ];then
+if [ ! $PROJECT = none ];then
     color_green $PROJECT
     sPROJECT=$color_word
 else
     sPROJECT=$PROJECT
 fi
 
-if [ $ENABLE_ALL == true ];then
+if [ $ENABLE_ALL = true ];then
     color_red $ENABLE_ALL
     sENABLE_ALL=$color_word
 else
@@ -393,7 +408,7 @@ echo "==========================="
 
 # main
 # create a tmp dir
-if [ ! -d "./pkg_tmp" ];then
+if [ ! -d ./pkg_tmp ];then
     mkdir ./pkg_tmp
 fi
 cd ./pkg_tmp
@@ -403,18 +418,27 @@ PARENT_PATH=$(pwd)
 
 # update 
 sudo apt update -y
-sudo apt upgrade -ys
+sudo apt upgrade -y
 
 # install development tools in ubuntu
 install_ubuntu_dev
 
 # install LRA project depend software and alias
 if [[ $PROJECT = "LRA" && $ONPI = true ]];then
-    echo "project is LRA"
+    color_red "project is LRA"
+    echo -e $color_word
     install_ubuntu_LRA_onpi
 fi
 
+sudo apt update -y
+sudo apt upgrade -y
+
 # if reboot flag on
-if [ $ENABLE_REBOOT == true ];then
+if [ $ENABLE_REBOOT = true ];then
+    echo "reboot starts"
     sudo reboot
+else
+    sudo su $USER
 fi
+
+
